refactor(checker): extract diagnostic emission methods from core.rs
Moves the diagnostic dedup, error(), push_diagnostic(), and message
normalization methods out of context/core.rs (1999 lines) into a new
dedicated module context/diagnostic_push.rs to stay within the 2000-line
hard limit (CLAUDE.md §19).

core.rs drops from 1999 → 1709 lines; diagnostic_push.rs is 299 lines.
No behavior change — the impl block split is transparent to callers.
