+++ crates/tsz-checker/src/context/mod.rs
@@ -24,6 +24,7 @@ mod constructors;
 mod core;
 mod cross_file_query;
 mod diagnostic_indices;
+mod diagnostic_push;
 mod env_eval_cache;
 mod file_session_reset;
 pub mod lifetime_shells;
